Reverted log_parser.py to what Nikhil had done, moving all yperf stuff into yperf_input.py and the yperf repository
